+++ academic.rb
@@ -1,6 +1,11 @@
 class Academic
   attr_reader :age, :phase
   attr_accessor :name
+  def initialize(options = {})
+    @age = options.fetch(:age, 0)
+    @name = options.fetch(:name, "")
+  end
+
   def offer_high_five
     "High five!"
   end
