+++ apprentice_teacher.rb
@@ -1,18 +1,9 @@
-require_relative 'academic'
-class ApprenticeTeacher < Academic
-  attr_reader :salary, :target_raise
-
+require_relative 'faculty'
+class ApprenticeTeacher < Faculty
 
   def initialize(options={})
-    @age = options.fetch(:age, 0)
-    @name = options.fetch(:name, "")
+    super
     @target_raise = 800
-    @phase = 3
-  end
-
-  def set_phase(num)
-    @phase = num
-    "Cool, I've always wanted to teach phase #{num}!"
   end
 
   def teach_stuff
@@ -21,15 +12,6 @@ class ApprenticeTeacher < Academic
     response += "*drops crazy knowledge bomb* "
     response += "... You're welcome."
     response
-  end
-
-  def salary=(new_salary)
-    puts "This better be good!"
-    @salary = new_salary
-  end
-
-  def receive_raise(raise)
-    @salary += raise
   end
 
   def set_performance_rating(rating)
@@ -42,6 +24,7 @@ class ApprenticeTeacher < Academic
     end
     response
   end
+
 
   def attend_training_session
     puts "Whoa. I know ruby-fu"
